+++ src/FREffects.hpp
@@ -1,0 +1,6 @@
+#include "FakeRate.hpp"
+
+class FREffects {
+public:
+    static CCSprite* grdInfinity();
+};